face-cut-function/requirements.txt
Pillow==9.0.1
boto3==1.24.70
ydb==2.10.0